+++ addons/yat/src/scenes/base_terminal/components/full_window_display/FullWindowDisplay.cs
@@ -1,0 +1,54 @@
+using Godot;
+using YAT.Helpers;
+
+namespace YAT.Scenes;
+
+public partial class FullWindowDisplay : Control
+{
+	[Signal] public delegate void OpenedEventHandler();
+	[Signal] public delegate void ClosedEventHandler();
+
+	public bool IsOpen => Visible;
+	public RichTextLabel MainDisplay { get; private set; }
+	private Label _helpLabel;
+
+	public override void _Ready()
+	{
+		MainDisplay = GetNode<RichTextLabel>("%MainDisplay");
+		_helpLabel = GetNode<Label>("%HelpLabel");
+	}
+
+	public override void _Input(InputEvent @event)
+	{
+		if (@event.IsActionPressed(Keybindings.TerminalCloseFullWindowDisplay))
+			Close();
+	}
+
+	public void Open(string text)
+	{
+		GenerateHelp();
+
+		MainDisplay.Clear();
+		MainDisplay.CallDeferred("append_text", text);
+		Visible = true;
+
+		EmitSignal(SignalName.Opened);
+	}
+
+	public void Close()
+	{
+		if (!Visible) return;
+
+		MainDisplay.Clear();
+		Visible = false;
+
+		EmitSignal(SignalName.Closed);
+	}
+
+	private void GenerateHelp()
+	{
+		var closeKey = InputMap.ActionGetEvents(Keybindings.TerminalCloseFullWindowDisplay)[0];
+
+		_helpLabel.Text = $"{closeKey.AsText()} - Close";
+	}
+}